+++ test/TST/ToPy.hs
@@ -6,6 +6,7 @@ import Data.ByteString      qualified as BS
 import Data.Set             qualified as Set
 import Data.Map.Strict      qualified as Map
 import Data.Complex         (Complex((:+)))
+import Numeric.Natural      (Natural)
 import Test.Tasty
 import Test.Tasty.HUnit
 import Python.Inline
@@ -57,4 +58,31 @@ tests = testGroup "ToPy"
   , testCase "dict unhashable" $ runPy $
       let x = Map.fromList [([1],10), ([5],50), ([3],30)] :: Map.Map [Int] Int
       in throwsPy [py_| x_hs |]
+    -- Integer. We only check large number here. Small will be well tested by QC
+  , testGroup "Integer" $ concat
+    [ [ testCase (" 2^"++show k++"-1") $ let n = 2^k - 1          :: Integer
+                                         in runPy [py_| assert n_hs == 2**k_hs - 1 |]
+      , testCase (" 2^"++show k)       $ let n = 2^k              :: Integer
+                                         in runPy [py_| assert n_hs == 2**k_hs |]
+      , testCase (" 2^"++show k++"+1") $ let n = 2^k + 1          :: Integer
+                                         in runPy [py_| assert n_hs == 2**k_hs + 1 |]
+      , testCase ("-2^"++show k++"-1") $ let n = negate $ 2^k - 1 :: Integer
+                                         in runPy [py_| assert n_hs == -(2**k_hs - 1) |]
+      , testCase ("-2^"++show k)       $ let n = negate $ 2^k     :: Integer
+                                         in runPy [py_| assert n_hs == -(2**k_hs) |]
+      , testCase ("-2^"++show k++"+1") $ let n = negate $ 2^k + 1 :: Integer
+                                         in runPy [py_|assert n_hs == -(2**k_hs + 1) |]
+      ]
+    | k <- [63,64,65,92,17,128,129,32100] :: [Int]
+    ]
+  , testGroup "Natural" $ concat
+    [ [ testCase (" 2^"++show k++"-1") $ let n = 2^k - 1          :: Natural
+                                         in runPy [py_| assert n_hs == 2**k_hs - 1 |]
+      , testCase (" 2^"++show k)       $ let n = 2^k              :: Natural
+                                         in runPy [py_| assert n_hs == 2**k_hs |]
+      , testCase (" 2^"++show k++"+1") $ let n = 2^k + 1          :: Natural
+                                         in runPy [py_| assert n_hs == 2**k_hs + 1 |]
+      ]
+    | k <- [63,64,65,92,17,128,129,32100] :: [Int]
+    ]
   ]